chore: separate `ActionButton` into their own file
